Changing structure

+++ sql_scripts/ddl/1_ddl.sql
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS plan
 (
     product_id INTEGER NOT NULL,
     shop_id INTEGER NOT NULL,
-    plan_cnt DECIMAL,
+    plan_cnt INTEGER,
     plan_date DATE,
     CONSTRAINT product_id FOREIGN KEY (product_id) REFERENCES products (product_id) ON DELETE CASCADE,
     CONSTRAINT shop_id FOREIGN KEY (shop_id) REFERENCES shops (shop_id) ON DELETE CASCADE
@@ -28,7 +28,7 @@ CREATE TABLE IF NOT EXISTS shop_citilink
     shop_id INTEGER NOT NULL,
     product_id INTEGER NOT NULL,
     sale_date DATE NOT NULL,
-    sales_cnt DECIMAL NOT NULL,
+    sales_cnt INTEGER NOT NULL,
     CONSTRAINT shop_id FOREIGN KEY (shop_id) REFERENCES shops (shop_id) ON DELETE CASCADE,
     CONSTRAINT product_id FOREIGN KEY (product_id) REFERENCES products (product_id) ON DELETE CASCADE
 );
@@ -38,7 +38,7 @@ CREATE TABLE IF NOT EXISTS shop_dns
     shop_id INTEGER NOT NULL,
     product_id INTEGER NOT NULL,
     sale_date DATE NOT NULL,
-    sales_cnt DECIMAL NOT NULL,
+    sales_cnt INTEGER NOT NULL,
     CONSTRAINT shop_id FOREIGN KEY (shop_id) REFERENCES shops (shop_id) ON DELETE CASCADE,
     CONSTRAINT product_id FOREIGN KEY (product_id) REFERENCES products (product_id) ON DELETE CASCADE
 );
@@ -48,7 +48,7 @@ CREATE TABLE IF NOT EXISTS shop_mvideo
     shop_id INTEGER NOT NULL,
     product_id INTEGER NOT NULL,
     sale_date DATE NOT NULL,
-    sales_cnt DECIMAL NOT NULL,
+    sales_cnt INTEGER NOT NULL,
     CONSTRAINT shop_id FOREIGN KEY (shop_id) REFERENCES shops (shop_id) ON DELETE CASCADE,
     CONSTRAINT product_id FOREIGN KEY (product_id) REFERENCES products (product_id) ON DELETE CASCADE
 );
